fix type

+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/process/last/LastQueryMergeOperator.java
@@ -300,7 +300,7 @@ public class LastQueryMergeOperator implements ProcessOperator {
 
   private static class Location {
 
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(Location.class).instanceSize();
+    private static final long INSTANCE_SIZE = ClassLayout.parseClass(Location.class).instanceSize();
     int tsBlockIndex;
     int rowIndex;
